+++ src/libopenrave/libopenrave.cpp
@@ -335,6 +335,7 @@ class RaveGlobal : private boost::noncopyable, public boost::enable_shared_from_
         _mapikparameterization[IKP_TranslationXY2D] = "TranslationXY2D";
         _mapikparameterization[IKP_TranslationXYOrientation3D] = "TranslationXYOrientation3D";
         _mapikparameterization[IKP_TranslationLocalGlobal6D] = "TranslationLocalGlobal6D";
+        _mapikparameterization[IKP_TranslationXAxisAngle4D] = "TranslationXAxisAngle4D";
         BOOST_ASSERT(_mapikparameterization.size()==IKP_NumberOfParameterizations);
     }
 public:
@@ -874,6 +875,11 @@ std::ostream& operator<<(std::ostream& O, const IkParameterization &ikparam)
         O << p.first.x << " " << p.first.y << " " << p.first.z << " " << p.second.x << " " << p.second.y << " " << p.second.z << " ";
         break;
     }
+    case IKP_TranslationXAxisAngle4D: {
+        std::pair<Vector,dReal> p = ikparam.GetTranslationXAxisAngle4D();
+        O << p.second << " " << p.first.x << " " << p.first.y << " " << p.first.z << " ";
+        break;
+    }
     default:
         throw OPENRAVE_EXCEPTION_FORMAT("does not support parameterization 0x%x", ikparam.GetType(),ORE_InvalidArguments);
     }
@@ -896,6 +902,7 @@ std::istream& operator>>(std::istream& I, IkParameterization& ikparam)
     case IKP_TranslationXY2D: { Vector v; I >> v.y >> v.y; ikparam.SetTranslationXY2D(v); break; }
     case IKP_TranslationXYOrientation3D: { Vector v; I >> v.y >> v.y >> v.z; ikparam.SetTranslationXYOrientation3D(v); break; }
     case IKP_TranslationLocalGlobal6D: { Vector localtrans, trans; I >> localtrans.x >> localtrans.y >> localtrans.z >> trans.x >> trans.y >> trans.z; ikparam.SetTranslationLocalGlobal6D(localtrans,trans); break; }
+    case IKP_TranslationXAxisAngle4D: { Vector trans; dReal angle=0; I >> trans.x >> trans.y >> trans.z >> angle; ikparam.SetTranslationXAxisAngle4D(trans,angle); break; }
     default:
         throw OPENRAVE_EXCEPTION_FORMAT("does not support parameterization 0x%x", ikparam.GetType(),ORE_InvalidArguments);
     }